First commit. Basic /order and /line apps.
    This is the first public commit of a project I've been building for the
past couple of weeks. Pizzeria is a point-of-sale system for pizza
restaurants. The basic structure is laid out for an ordering system, as well
as a line system.
    The order app has a barebones interface for adding pizzas and breads to an
order. It's not exactly robust, as an empty order is created with each new
visit. This order stays in the system, as there's no concept yet of "Sending"
an order to the line. Instead, temporary objects must be constructed for the
order, and the order must be deleted if not sent. This action has to take
place when the user leaves the order app, so I'm not sure how to implement
this yet. I'm still learning as I go.
    The line app is a recent addition, querying the orders for those not yet
cooked (based on a BooleanField). It currently lacks a way to flip the
"is_done" flag, but that should be easy enough. I'm thinking I should make
each order in the line page its own form, which upon submission will flip the
is_done field to True.
    There are some basic model tests for the order app, although the one view
test I constructed fails in search of a 404 template. After discovery of a
related Stack Overflow question <http://bit.ly/h2W7Oa>, I think I have to
either monkeypatch as suggested on SO, or possibly use a Mock, at the
suggestion of Ned Batchelder <http://blip.tv/file/3261272/>.
    Before I touch the working views, however, I need to fix the price fields
in the models to be more uniform and, well, usable. At first I had intended to
separate the price information from the models describing the items, but this
seems like a bad idea. I have to figure out how to make the models self-aware
of their price, and scale it according to size, toppings, type, etc. depending
on the item.
    What began as an experiment in Django has become a full-fledged project,
and I'm having a blast. I'll be picking out the experimental parts left over,
and adding to the site's functionality. The real fun comes when I get to use
Javascript, and make things really spiffy :). In addition to re-learning the
web stack up and down, of course. 'Til next time.