Basic line functionality, buttons to clear orders
    I've implemented a basic mechanism to clear orders in the line app, using
forms. There was no need to use Django's Form models, as there's no data to
validate. The only relevant data is the button used to submit the form, passed
through the POST dictionary. However, the buttons and the page in general
reflect system-wide order numbers, rather than the unmade orders I'm concerned
with here. The next step is to get those unmade orders, and assign them a new
number, ordered by the time made (first in, first out). The trick is, to do
this with a query to the Order object, or find out if I can pass a precomputed
QuerySet to the generic view.