Base prices functional, Order updates not working
    The recipt now reflects price changes, or at least tries to. The base
prices are showing fine, but the total/subtotal/tax fields are not. I think
when the pizza is updated, the order may not be. This would keep the fields
from changing as they should. Also, the price spans in the template aren't
showing inline, but I really didn't expect them to by default, as they're
sitting next to a p block. This is a secondary concern. Right now, I have to
figure out why my order isn't updating.