Prices fully functional, decimal issues solved
    The behavior of the decimal fields has been corrected, by two methods.
First, I removed the manual override on decimal precision. Precision in the
decimal module is not the number of digits after the decimal, as I thought,
but the number of digits in the whole number. After sifting through pages on
Stack Overflow, I found a better solution. I could keep the default precision,
do my calculations, then save the result with a Decimal.quantize() filter. The
quantize function takes a Decimal, as in '0.01' for my format, and returns the
Decimal with the proper number of decimal places.
    Once this was fixed, the problem of trailing zeros remained. Luckily, I
knew where to find this: template tags. With the '|stringformat:".2f"' filter,
the prices are piped through a familiar string format. The prices have stopped
their strange rounding, and are displaying properly. Now, I think I should
separate the price logic, into a Register model. This price aggregation
function will be needed, among others, throughout the apps.